feat!: use map of objects for dynamic map blocks (#45)
* feat!: use map of objects for dynamic map blocks

Instead of using a list of objects, we use now a map of objects.
This would make our declared volumes in our module using a key instead
of a index.

The problem with the key is that it is order dependent. So if the
volumes order is switched in our container module, it makes our volumes
to be deleted & recreated. By referencing it on a key string the order
should not matter.

Fixes #44

BREAKING CHANGE: syntax for volume declaration has changed.

Before:
```
named_volumes = [
  {
    container_name = "nginx_config"
    container_path = "/etc/nginx"
    read_only      = false
    create         = true
  }
]
host_paths = [
  {
    host_path      = "/media/local"
    container_path = "/mnt"
    read_only      = true
  }
]
volumes_from_containers = [
  {
    container_name = "proxy"
  }
]
devices = [
  {
    host_path      = "/dev/vcs"
    container_path = "/dev/vcs"
    permissions    = "rwm"
  }
]
docker_networks = [
  {
    name = "proxy-tier"
    ipam_config = {
      aux_address = {}
      gateway     = "10.0.30.1"
      subnet      = "10.0.30.0/24"
    }
  }
]
```

Now:
```
named_volumes = {
  "nginx_config" = {
    container_path = "/etc/nginx"
    read_only      = false
    create         = true
  }
}
host_paths = {
  "/media/local" = {
    container_path = "/mnt"
    read_only      = true
  }
}
volumes_from_containers = [
  "proxy"
]
devices = {
  "/dev/vcs" = {
    container_path = "/dev/vcs"
    permissions    = "rwm"
  }
}
docker_networks = {
  "proxy-tier" = {
    ipam_config = {
      aux_address = {}
      gateway     = "10.0.30.1"
      subnet      = "10.0.30.1/24"
    }
  }
}
```

* chore: docker provider has moved to kreuzwerker/docker